Remove non-safe cast from int to int32

+++ python/dolfinx/mesh.py
@@ -203,7 +203,7 @@ def MeshTags(mesh: Mesh, dim: int, indices: np.ndarray, values: np.ndarray) -> t
     """
 
     if isinstance(values, int):
-        values = np.full(indices.shape, values, dtype=np.int32)
+        values = np.full(indices.shape, values, dtype=np.int64)
     elif isinstance(values, float):
         values = np.full(indices.shape, values, dtype=np.double)
